Change behaviour towards modified files (selective cleanup)
Before this commit, when a file was modified, the plugin
would clean all previous outputs for the file.
It now only cleans stale icons (icons that are not present
after the modification of the file), avoiding some IO
deletion operation on the filesystem of icons that would
be re-created/updated anyway.

This commit also cleans up a bit the parsing and state logic.