Compact state file content + relativize state file names to make cache sharable in the best case
The state file name is relative to the [projectDirectory] to make state file agnostic of the developer's directory
structure before the project directory. This makes the output cache of the plugin sharable amongst the different
machines (CI, developers laptop, etc.) where the plugin used.

In the case the plugin is used on sources that are before the project directory, the cache may or may not be sharable.
For this reason, it is discouraged to use the plugin on files outside the project directory. The cache will still
work as expected, state file names will simply include dots.